refactor: refatora arquitetura, melhora de tratamento das exceptions, adiciona paginacao, seeder admin e rota /usuarios

+++ src/main/java/br/com/barros/Cinema/dto/SessaoResponseDto.java
@@ -1,11 +1,12 @@
 package br.com.barros.Cinema.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record SessaoResponseDto (
         Long id,
         LocalDateTime hora,
         Integer sala,
-        Double preco,
+        BigDecimal preco,
         Long filmeId
 ){}
